requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
akshare==1.16.87
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fonttools==4.57.0
frozenlist==1.6.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
jiter==0.9.0
joblib==1.5.0
jsonpath==0.82.2
kiwisolver==1.4.7
lxml==5.4.0
matplotlib==3.9.4
mini-racer==0.12.4
multidict==6.4.3
nest-asyncio==1.6.0
numpy==2.0.2
openai==1.77.0
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
ta==0.11.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
xlrd==2.0.1
yarl==1.20.0
zipp==3.21.0
